Unify commands
